Add files via upload
Prototype of a recipe app that is supposed to tell you exact amount of each ingredient to buy so you dont end up wasting any food. Also generates a shopping list for the store that is closest, or another store if it has a better deal. user can choose if they want to go to multiple stores if there are different deals at other stores.